refactor GithubQuickStartProject to its own file
